pyproject.toml
[project]
name = "session-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.39.0",
    "markdownify>=1.2.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "streamlit>=1.50.0",
    "tavily-python>=0.7.12",
]
